+++ src/image.js
@@ -1,6 +1,7 @@
 import { pipe, withConstructor, withStaticProperty } from '@flekschas/utils';
 import * as PIXI from 'pixi.js';
 
+import withClone from './with-clone';
 import withColorFilters from './with-color-filters';
 import withDestroy from './with-destroy';
 import withScale from './with-scale';
@@ -20,13 +21,14 @@ const createImage = (source, { anchor = [0.5, 0.5] } = {}) => {
   }
 
   return pipe(
+    withConstructor(createImage),
     withStaticProperty('displayObject', sprite),
     withStaticProperty('sprite', sprite),
+    withClone(displayObject, { anchor }),
     withColorFilters(sprite),
     withScale(sprite, displayObject.width, displayObject.height),
     withSize(sprite, displayObject.width, displayObject.height),
-    withDestroy(sprite),
-    withConstructor(createImage)
+    withDestroy(sprite)
   )({});
 };